Fix ds2438 readout and Humidity format

+++ src/digitemp.h
@@ -90,7 +90,7 @@ void usage();
 void free_coupler();
 float c2f( float temp );
 int build_tf( char *time_format, char *format, int sensor, 
-              float temp_c, int humidity, unsigned char *sn );
+              float temp_c, float humidity, unsigned char *sn );
 int build_cf( char *time_format, char *format, int sensor, int page,
               unsigned long count, unsigned char *sn );
 int build_af(char *time_format, size_t tf_size, char *format,
@@ -99,7 +99,7 @@ int build_af(char *time_format, size_t tf_size, char *format,
 int log_string( char *line );
 int log_temp( int sensor, float temp_c, unsigned char *sn );
 int log_counter( int sensor, int page, unsigned long counter, unsigned char *sn );
-int log_humidity( int sensor, double temp_c, int humidity, unsigned char *sn );
+int log_humidity( int sensor, double temp_c, float humidity, unsigned char *sn );
 int log_temperature_voltage(int sensor, double temp_c,
                             float vdd, float ad, float vsens,
                             unsigned char *sn);
